+++ tests/testthat/test-03-EngineRestGitHub.R
@@ -140,7 +140,8 @@ test_that("`pull_repos_issues()` adds issues to repos table", {
 test_that("`pull_repos_contributors()` adds contributors to repos table", {
   expect_snapshot(
     gh_repos_by_phrase_table <- test_rest$pull_repos_contributors(
-      test_mocker$use("gh_repos_by_phrase_table")
+      test_mocker$use("gh_repos_by_phrase_table"),
+      settings = test_settings
     )
   )
   expect_repos_table(